improved generation_and_dissipation_terms and renamed file
